move much Akka code from Example to Adapter
