+++ examples/src/main/scala/caliban/akkahttp/ExampleApp.scala
@@ -1,18 +1,9 @@
 package caliban.akkahttp
 
-import akka.NotUsed
-
 import scala.io.StdIn
-import akka.actor.{ ActorRef, ActorSystem }
-import akka.stream.CompletionStrategy
-
+import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
-import akka.http.scaladsl.model.ws.{ BinaryMessage, Message, TextMessage }
 import akka.http.scaladsl.server.Directives._
-import akka.http.scaladsl.model.ws.TextMessage.Streamed
-
-import akka.stream.OverflowStrategy
-import akka.stream.scaladsl.{ Flow, Keep, Sink, Source }
 import caliban.ExampleData.{ sampleCharacters, Character, CharacterArgs, CharactersArgs, Role }
 import caliban.GraphQL.graphQL
 import caliban.schema.Annotations.{ GQLDeprecated, GQLDescription }
@@ -22,8 +13,6 @@ import zio.clock.Clock
 import zio.console.Console
 import zio.stream.ZStream
 import zio.{ DefaultRuntime, URIO }
-
-import scala.concurrent.Future
 
 object ExampleApp extends App with GenericSchema[Console with Clock] {
 
@@ -35,59 +24,6 @@ object ExampleApp extends App with GenericSchema[Console with Clock] {
   implicit val characterSchema      = gen[Character]
   implicit val characterArgsSchema  = gen[CharacterArgs]
   implicit val charactersArgsSchema = gen[CharactersArgs]
-
-  trait Protocol
-  case object Complete           extends Protocol
-  case class Fail(ex: Exception) extends Protocol
-
-  def handleMessage(message: String, actor: ActorRef): Unit =
-    println(s"handleMessage got ws message '$message' from client")
-  // should hook up to AkkaHttpAdapter to act on the subscription
-
-  def createFlow(actor: ActorRef): Sink[Message, NotUsed] =
-    Flow[Message].map {
-      case TextMessage.Strict(message) =>
-        println(s"createFlow got message $message")
-        handleMessage(message, actor)
-        Nil
-      case Streamed(textStream) =>
-        textStream.runFold("")(_ + _).map(message => handleMessage(message, actor)).flatMap(Future.successful)
-        Nil
-      case bm: BinaryMessage =>
-        println(s"createFlow got BinaryMessage ")
-        // ignore binary messages but drain content to avoid the stream being clogged
-        bm.dataStream.runWith(Sink.ignore)
-        Nil
-    }.to(Sink.ignore)
-
-  def createSource(): (Source[TextMessage.Strict, NotUsed], ActorRef) = {
-    println("createSource")
-    val completionMatcher: PartialFunction[Any, CompletionStrategy] = {
-      case Complete =>
-        println("createSource got Complete")
-        CompletionStrategy.draining
-      case Fail(ex) =>
-        println(s"createSource got Fail ${ex}")
-        CompletionStrategy.draining
-    }
-    val failureMatcher: PartialFunction[Any, Throwable] = {
-      case Fail(ex) =>
-        println(s"createSource failureMatcher ${ex}")
-        ex
-    }
-    val (ref, publisher) =
-      Source
-        .actorRef[TextMessage.Strict](
-          completionMatcher,
-          failureMatcher,
-          bufferSize = Int.MaxValue,
-          overflowStrategy = OverflowStrategy.fail
-        )
-        .toMat(Sink.asPublisher(true))(Keep.both)
-        .run()
-
-    (Source.fromPublisher(publisher), ref)
-  }
 
   case class Queries(
     @GQLDescription("Return all characters from a given origin")
@@ -127,13 +63,7 @@ object ExampleApp extends App with GenericSchema[Console with Clock] {
     } ~ path("graphiql") {
       getFromResource("graphiql.html")
     } ~ path("ws" / "graphql") {
-      get {
-        val (source, actor)              = createSource()
-        val sink: Sink[Message, NotUsed] = createFlow(actor)
-        extractUpgradeToWebSocket { upgrade ⇒
-          complete(upgrade.handleMessagesWithSinkSource(sink, source, subprotocol = Some("graphql-ws")))
-        }
-      }
+      AkkaHttpAdapter.makeWebSocketService(interpreter)
     }
 
   val bindingFuture = Http().bindAndHandle(route, "localhost", 8088)
